initial device removal
